test(hub): split read-surface and cache tests into subject files
